v0.18~preview.130.74+175

+++ avx/int16x8.ml
@@ -1,7 +1,7 @@
 include Ocaml_simd_sse.Int16x8
 
 external low_of
-  :  int64#
+  :  int16#
   -> t
   @@ portable
   = "ocaml_simd_avx_unreachable" "caml_int16x8_low_of_int"
